plumbing: transport, Remove duplicate tests
