+++ plumbing/transport/http/upload_pack_test.go
@@ -49,23 +49,6 @@ func (s *UploadPackSuite) TestAdvertisedReferencesNotExists() {
 	s.Nil(conn)
 }
 
-// func (s *UploadPackSuite) TestuploadPackRequestToReader() {
-// 	r := &transport.FetchRequest{}
-// 	r.Wants = append(r.Wants, plumbing.NewHash("d82f291cde9987322c8a0c81a325e1ba6159684c"))
-// 	r.Wants = append(r.Wants, plumbing.NewHash("2b41ef280fdb67a9b250678686a0c3e03b0a9989"))
-// 	r.Haves = append(r.Haves, plumbing.NewHash("6ecf0ef2c2dffb796033e5a02219af86ec6584e5"))
-//
-// 	sr, err := uploadPackRequestToReader(r)
-// 	s.Nil(err)
-// 	b, _ := io.ReadAll(sr)
-// 	s.Equal(string(b),
-// 		"0032want 2b41ef280fdb67a9b250678686a0c3e03b0a9989\n"+
-// 			"0032want d82f291cde9987322c8a0c81a325e1ba6159684c\n0000"+
-// 			"0032have 6ecf0ef2c2dffb796033e5a02219af86ec6584e5\n"+
-// 			"0009done\n",
-// 	)
-// }
-
 func (s *UploadPackSuite) TestAdvertisedReferencesRedirectPath() {
 	endpoint, _ := transport.NewEndpoint("https://gitlab.com/gitlab-org/gitter/webapp")
