+++ plumbing/transport/ssh/proxy_test.go
@@ -4,31 +4,43 @@ import (
 	"context"
 	"fmt"
 	"net"
-	"sync/atomic"
+	"testing"
 
 	"github.com/armon/go-socks5"
-	fixtures "github.com/go-git/go-git-fixtures/v5"
 	ttest "github.com/go-git/go-git/v6/internal/transport/test"
 	"github.com/go-git/go-git/v6/plumbing/transport"
-	"github.com/go-git/go-git/v6/storage/filesystem"
+	"github.com/stretchr/testify/suite"
 
 	stdssh "golang.org/x/crypto/ssh"
 )
 
 type ProxySuite struct {
-	UploadPackSuite
+	suite.Suite
 }
 
-var socksProxiedRequests int32
+func TestProxySuite(t *testing.T) {
+	suite.Run(t, new(ProxySuite))
+}
 
+type TestProxyRule struct{ proxiedRequests int }
+
+func (dr *TestProxyRule) Allow(ctx context.Context, req *socks5.Request) (context.Context, bool) {
+	dr.proxiedRequests++
+	return ctx, true
+}
+
+// This test tests proxy support via an env var, i.e. `ALL_PROXY`.
+// Its located in a separate package because golang caches the value
+// of proxy env vars leading to misleading/unexpected test results.
 func (s *ProxySuite) TestCommand() {
 	socksListener := ttest.ListenTCP(s.T())
 
+	rule := TestProxyRule{}
 	socksServer, err := socks5.New(&socks5.Config{
 		AuthMethods: []socks5.Authenticator{socks5.UserPassAuthenticator{
 			Credentials: socks5.StaticCredentials{"user": "pass"},
 		}},
-		Rules: TestProxyRule{},
+		Rules: &rule,
 	})
 	s.Require().NoError(err)
 
@@ -38,16 +50,13 @@ func (s *ProxySuite) TestCommand() {
 
 	socksProxyAddr := fmt.Sprintf("socks5://localhost:%d", socksListener.Addr().(*net.TCPAddr).Port)
 
-	s.base, s.port, _ = setupTest(s.T())
+	base, port, _ := setupTest(s.T())
 
 	DefaultAuthBuilder = func(user string) (AuthMethod, error) {
 		return &Password{User: user}, nil
 	}
 
-	ep := newEndpoint(s.T(), s.base, s.port, "basic.git")
-	basic := ttest.PrepareRepository(s.T(), fixtures.Basic().One(), s.base, "basic.git")
-	s.Storer = filesystem.NewStorage(basic, nil)
-	s.NoError(err)
+	ep := newEndpoint(s.T(), base, port, "")
 	ep.Proxy = transport.ProxyOptions{
 		URL:      socksProxyAddr,
 		Username: "user",
@@ -61,13 +70,6 @@ func (s *ProxySuite) TestCommand() {
 	}
 	_, err = runner.Command(context.TODO(), transport.UploadPackService.String(), ep, nil)
 	s.NoError(err)
-	proxyUsed := atomic.LoadInt32(&socksProxiedRequests) > 0
-	s.True(proxyUsed)
-}
 
-type TestProxyRule struct{}
-
-func (dr TestProxyRule) Allow(ctx context.Context, req *socks5.Request) (context.Context, bool) {
-	atomic.AddInt32(&socksProxiedRequests, 1)
-	return ctx, true
+	s.True(rule.proxiedRequests > 0)
 }
